Add new test based on config provided by Anna here: https://issues.jboss.org/browse/ISPN-3432
The config was fixed so that each of the two nodes uses a different jdbc database name for the (not shared) cache store of the infinispan dir internal caches. After this change is done the test no longer fails. See differences between initial file and async-config.xml and async-config2.xml.